Use old style test for survey response update tests
